Feat: Adjusting entities.

+++ L.GastosProdutos.Core/Domain/Entities/Product/ProductEntity.cs
@@ -4,7 +4,7 @@ namespace L.GastosProdutos.Core.Domain.Entities.Product
 {
     public class ProductEntity : BaseEntity
     {
-        public ProductEntity(string name, decimal price, float quantity)
+        public ProductEntity(string name, decimal price, decimal quantity)
         {
             Name = name;
             Price = price;
@@ -15,11 +15,11 @@ namespace L.GastosProdutos.Core.Domain.Entities.Product
 
         public decimal Price { get; set; }
 
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         public decimal UnitPrice => GetUnitPrice();
 
         private decimal GetUnitPrice() =>
-            Price / (decimal)Quantity;
+            Price / Quantity;
     }
 }
